requirements.txt
click
rich
GitPython
toml